+++ libconsensus/pbft/Common.h
@@ -1,0 +1,415 @@
+/*
+ * @CopyRight:
+ * FISCO-BCOS is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * FISCO-BCOS is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with FISCO-BCOS.  If not, see <http://www.gnu.org/licenses/>
+ * (c) 2016-2018 fisco-dev contributors.
+ */
+
+/**
+ * @brief Common functions and types of Consensus module
+ * @author: yujiechen
+ * @date: 2018-09-21
+ */
+#pragma once
+#include <libconsensus/Common.h>
+#include <libdevcore/RLP.h>
+#include <libdevcore/SHA3.h>
+#include <libdevcrypto/Common.h>
+#include <libethcore/Block.h>
+#include <libethcore/Exceptions.h>
+namespace dev
+{
+namespace consensus
+{
+DEV_SIMPLE_EXCEPTION(DisabledFutureTime);
+DEV_SIMPLE_EXCEPTION(InvalidBlockHeight);
+DEV_SIMPLE_EXCEPTION(ExistedBlock);
+DEV_SIMPLE_EXCEPTION(ParentNoneExist);
+DEV_SIMPLE_EXCEPTION(BlockMinerListWrong);
+
+// for pbft
+enum PBFTPacketType : byte
+{
+    PrepareReqPacket = 0x00,
+    SignReqPacket = 0x01,
+    CommitReqPacket = 0x02,
+    ViewChangeReqPacket = 0x03,
+    PBFTPacketCount
+};
+
+/// PBFT message
+struct PBFTMsgPacket
+{
+    /// the index of the node that sends this pbft message
+    u256 node_idx;
+    /// the node id of the node that sends this pbft message
+    h512 node_id;
+    /// type of the packet(maybe prepare, sign or commit)
+    /// (receive from the network or send to the network)
+    unsigned packet_id;
+    /// the data of concrete request(receive from or send to the network)
+    bytes data;
+    /// timestamp of receive this pbft message
+    u256 timestamp;
+    /// default constructor
+    PBFTMsgPacket() : node_idx(h256(0)), node_id(h512(0)), packet_id(0), timestamp(u256(utcTime()))
+    {}
+
+    bool operator==(PBFTMsgPacket const& msg)
+    {
+        return node_idx == msg.node_idx && node_id == msg.node_id && packet_id == msg.packet_id &&
+               data == msg.data && timestamp == msg.timestamp;
+    }
+    bool operator!=(PBFTMsgPacket const& msg) { return !operator==(msg); }
+    /**
+     * @brief : encode network-send part of PBFTMsgPacket into bytes (RLP encoder)
+     * @param encodedBytes: encoded bytes of the network-send part of PBFTMsgPacket
+     */
+    virtual void encode(bytes& encodedBytes) const
+    {
+        RLPStream tmp;
+        streamRLPFields(tmp);
+        tmp.swapOut(encodedBytes);
+    }
+    /**
+     * @brief : decode the network-receive part of PBFTMsgPacket into PBFTMsgPacket object
+     * @param data: network-receive part of PBFTMsgPacket
+     * @ Exception Case: if decode failed, we throw exceptions
+     */
+    virtual void decode(bytesConstRef data) { populate(data); }
+
+    /// RLP decode: serialize network-received packet-data from bytes to RLP
+    void streamRLPFields(RLPStream& s) const { s.append(unsigned(packet_id)).append(data); }
+
+    /**
+     * @brief: set non-network-receive-or-send part of PBFTMsgPacket
+     * @param idx: the index of the node that send the PBFTMsgPacket
+     * @param nodeId : the id of the node that send the PBFTMsgPacket
+     */
+    void setOtherField(u256 const& idx, h512 const& nodeId)
+    {
+        node_idx = idx;
+        node_id = nodeId;
+        timestamp = u256(utcTime());
+    }
+    /// populate PBFTMsgPacket from RLP object
+    void populate(bytesConstRef req_data)
+    {
+        try
+        {
+            packet_id = (unsigned)RLP(req_data.cropped(0, 1)).toInt<unsigned>(RLP::ThrowOnFail);
+            data = RLP(req_data.cropped(1)).toBytes();
+        }
+        catch (Exception const& e)
+        {
+            e << dev::eth::errinfo_name("invalid msg format");
+            throw;
+        }
+    }
+};
+
+/// the base class of PBFT message
+struct PBFTMsg
+{
+    /// the number of the block that is handling
+    int64_t height = -1;
+    /// view when construct this PBFTMsg
+    u256 view = Invalid256;
+    /// index of the node generate the PBFTMsg
+    u256 idx = Invalid256;
+    /// timestamp when generate the PBFTMsg
+    u256 timestamp = Invalid256;
+    /// block-header hash of the block handling
+    h256 block_hash = h256();
+    /// signature to the block_hash
+    Signature sig = Signature();
+    /// signature to the hash of other fields except block_hash, sig and sig2
+    Signature sig2 = Signature();
+    PBFTMsg() = default;
+    PBFTMsg(KeyPair const& _keyPair, int64_t const& _height, u256 const& _view, u256 const& _idx,
+        h256 const _blockHash)
+    {
+        height = _height;
+        view = _view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = _blockHash;
+        sig = signHash(block_hash, _keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), _keyPair);
+    }
+
+    bool operator==(PBFTMsg const& req)
+    {
+        return height == req.height && view == req.view && timestamp == req.timestamp &&
+               block_hash == req.block_hash && sig == req.sig && sig2 == req.sig2;
+    }
+
+    bool operator!=(PBFTMsg const& req) { return !operator==(req); }
+    /**
+     * @brief: encode the PBFTMsg into bytes
+     * @param encodedBytes: the encoded bytes of specified PBFTMsg
+     */
+    virtual void encode(bytes& encodedBytes) const
+    {
+        RLPStream tmp;
+        streamRLPFields(tmp);
+        RLPStream list_rlp;
+        list_rlp.appendList(1).append(tmp.out());
+        list_rlp.swapOut(encodedBytes);
+    }
+
+    /**
+     * @brief : decode the bytes received from network into PBFTMsg object
+     * @param data : network-received data to be decoded
+     * @param index: the index of RLP data need to be populated
+     * @Exception Case: if decode failed, throw exception directly
+     */
+    virtual void decode(bytesConstRef data, size_t const& index = 0)
+    {
+        RLP rlp(data);
+        populate(rlp[index]);
+    }
+
+    /// trans PBFTMsg into RLPStream for encoding
+    virtual void streamRLPFields(RLPStream& _s) const
+    {
+        _s << height << view << idx << timestamp << block_hash << sig.asBytes() << sig2.asBytes();
+    }
+
+    /// populate specified rlp into PBFTMsg object
+    virtual void populate(RLP const& rlp)
+    {
+        int field = 0;
+        try
+        {
+            height = rlp[field = 0].toInt<int64_t>();
+            view = rlp[field = 1].toInt<u256>();
+            idx = rlp[field = 2].toInt<u256>();
+            timestamp = rlp[field = 3].toInt<u256>();
+            block_hash = rlp[field = 4].toHash<h256>(RLP::VeryStrict);
+            sig = dev::Signature(rlp[field = 5].toBytesConstRef());
+            sig2 = dev::Signature(rlp[field = 6].toBytesConstRef());
+        }
+        catch (Exception const& _e)
+        {
+            _e << dev::eth::errinfo_name("invalid msg format")
+               << dev::eth::BadFieldError(field, toHex(rlp[field].data().toBytes()));
+            throw;
+        }
+    }
+
+    /// clear the PBFTMsg
+    void clear()
+    {
+        height = -1;
+        view = Invalid256;
+        idx = Invalid256;
+        timestamp = Invalid256;
+        block_hash = h256();
+        sig = Signature();
+        sig2 = Signature();
+    }
+
+    /// get the hash of the fields without block_hash, sig and sig2
+    h256 fieldsWithoutBlock() const
+    {
+        RLPStream ts;
+        ts << height << view << idx << timestamp;
+        return dev::sha3(ts.out());
+    }
+
+    /**
+     * @brief : sign for specified hash using given keyPair
+     * @param hash: hash data need to be signed
+     * @param keyPair: keypair used to sign for the specified hash
+     * @return Signature: signature result
+     */
+    Signature signHash(h256 const& hash, KeyPair const& keyPair) const
+    {
+        return dev::sign(keyPair.secret(), hash);
+    }
+};
+
+/// definition of the prepare requests
+struct PrepareReq : public PBFTMsg
+{
+    /// block data
+    bytes block;
+    /// execution result of block(save the execution result temporarily)
+    /// no need to send or receive accross the network
+    dev::blockverifier::ExecutiveContext::Ptr p_execContext;
+    /// default constructor
+    PrepareReq() = default;
+    PrepareReq(KeyPair const& _keyPair, int64_t const& _height, u256 const& _view, u256 const& _idx,
+        h256 const _blockHash)
+      : PBFTMsg(_keyPair, _height, _view, _idx, _blockHash), p_execContext(nullptr)
+    {}
+    /**
+     * @brief: populate the prepare request from specified prepare request,
+     *         given view and node index
+     *
+     * @param req: given prepare request to populate the PrepareReq object
+     * @param keyPair: keypair used to sign for the PrepareReq
+     * @param _view: current view
+     * @param _idx: index of the node that generates this PrepareReq
+     */
+    PrepareReq(PrepareReq const& req, KeyPair const& keyPair, u256 const& _view, u256 const& _idx)
+    {
+        height = req.height;
+        view = _view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = req.block_hash;
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+        block = req.block;
+        p_execContext = nullptr;
+    }
+
+    /**
+     * @brief: construct PrepareReq from given block, view and node idx
+     * @param blockStruct : the given block used to populate the PrepareReq
+     * @param keyPair : keypair used to sign for the PrepareReq
+     * @param _view : current view
+     * @param _idx : index of the node that generates this PrepareReq
+     */
+    PrepareReq(
+        dev::eth::Block& blockStruct, KeyPair const& keyPair, u256 const& _view, u256 const& _idx)
+    {
+        height = blockStruct.blockHeader().number();
+        view = _view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = blockStruct.blockHeader().hash();
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+        blockStruct.encode(block);
+        p_execContext = nullptr;
+    }
+
+    /**
+     * @brief : update the PrepareReq with specified block and block-execution-result
+     *
+     * @param sealing : object contains both block and block-execution-result
+     * @param keyPair : keypair used to sign for the PrepareReq
+     */
+    void updatePrepareReq(Sealing& sealing, KeyPair const& keyPair)
+    {
+        timestamp = u256(utcTime());
+        block_hash = sealing.block.blockHeader().hash();
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+        sealing.block.encode(block);
+        p_execContext = sealing.p_execContext;
+        LOG(DEBUG) << "Re-generate prepare_requests since block has been executed, time = "
+                   << timestamp;
+    }
+
+    /// trans PrepareReq from object to RLPStream
+    virtual void streamRLPFields(RLPStream& _s) const
+    {
+        PBFTMsg::streamRLPFields(_s);
+        _s << block;
+    }
+
+    /// populate PrepareReq from given RLP object
+    virtual void populate(RLP const& _rlp)
+    {
+        PBFTMsg::populate(_rlp);
+        int field = 0;
+        try
+        {
+            block = _rlp[field = 7].toBytes();
+        }
+        catch (Exception const& _e)
+        {
+            _e << dev::eth::errinfo_name("invalid msg format")
+               << dev::eth::BadFieldError(field, toHex(_rlp[field].data().toBytes()));
+            throw;
+        }
+    }
+};
+
+/// signature request
+struct SignReq : public PBFTMsg
+{
+    SignReq() = default;
+
+    /**
+     * @brief: populate the SignReq from given PrepareReq and node index
+     *
+     * @param req: PrepareReq used to populate the SignReq
+     * @param keyPair: keypair used to sign for the SignReq
+     * @param _idx: index of the node that generates this SignReq
+     */
+    SignReq(PrepareReq const& req, KeyPair const& keyPair, u256 const& _idx)
+    {
+        height = req.height;
+        view = req.view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = req.block_hash;
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+    }
+};
+
+/// commit request
+struct CommitReq : public PBFTMsg
+{
+    CommitReq() = default;
+    /**
+     * @brief: populate the CommitReq from given PrepareReq and node index
+     *
+     * @param req: PrepareReq used to populate the CommitReq
+     * @param keyPair: keypair used to sign for the CommitReq
+     * @param _idx: index of the node that generates this CommitReq
+     */
+    CommitReq(PrepareReq const& req, KeyPair const& keyPair, u256 const& _idx)
+    {
+        height = req.height;
+        view = req.view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = req.block_hash;
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+    }
+};
+
+/// view change request
+struct ViewChangeReq : public PBFTMsg
+{
+    ViewChangeReq() = default;
+    /**
+     * @brief: generate ViewChangeReq from given params
+     * @param keyPair: keypair used to sign for the ViewChangeReq
+     * @param _height: current block number
+     * @param _view: current view
+     * @param _idx: index of the node that generates this ViewChangeReq
+     * @param _hash: block hash
+     */
+    ViewChangeReq(KeyPair const& keyPair, uint64_t const& _height, u256 const _view,
+        u256 const& _idx, h256 const& _hash)
+    {
+        height = _height;
+        view = _view;
+        idx = _idx;
+        timestamp = u256(utcTime());
+        block_hash = _hash;
+        sig = signHash(block_hash, keyPair);
+        sig2 = signHash(fieldsWithoutBlock(), keyPair);
+    }
+};
+}  // namespace consensus
+}  // namespace dev
